refactor: move wallet_canister_id to dfx-core
