+++ src/dfx/src/lib/migrate.rs
@@ -1,9 +1,9 @@
-use crate::lib::identity::wallet::wallet_canister_id;
 use crate::lib::operations::canister::install_wallet;
 use crate::lib::{environment::Environment, error::DfxResult, root_key::fetch_root_key_if_needed};
 use anyhow::{bail, Context, Error};
 use candid::{CandidType, Deserialize, Principal};
 use dfx_core::config::model::network_descriptor::NetworkDescriptor;
+use dfx_core::identity::wallet::wallet_canister_id;
 use dfx_core::identity::Identity;
 use ic_agent::{Agent, Identity as _};
 use ic_utils::{
